[NVVM][MLIR] Refactor conversion of Math / Arith Operations seperate Passes
This Commit refactors the conversion of Math / Arith operations to NVVM into
a separate Pass called MathToNVVM. This was done to allow to support the
lowering of Math / Arith operations in flang. This mirrors what was done
in MathToROCDL.
